moved all something structs together
